refactor(services): extract Q&A generation into qa_generation_service.py (RFC #2)
Moves generate_answer_drafts, generate_answer_drafts_cascade,
generate_answer_drafts_parallel, generate_answer_drafts_multi, and all
their helpers (_PROVIDER_RANK, _extract_category_framework,
_parse_answer_drafts, _rule_based_answer_drafts,
_rule_based_cover_letter_drafts) out of resume_generator.py into the
dedicated qa_generation_service.py.

vault/answers.py now imports directly from qa_generation_service.
resume_generator.py retains thin re-export shims (clearly marked RFC #2)
so no other call sites break in this PR — shims will be removed in RFC #3.

Zero behaviour change: 48 unit tests pass, mypy clean, ruff+black clean.
